Viewer: revert R8 quantization; render at full Float32 precision
757e85f traded visual fidelity for a 4x texture-upload win. Smoke
gradients showed visible stepping at low concentrations (5-50 ug/m³
spent only ~25 of the 256 quantization levels), making distinct EPA
category 'patches' instead of continuous gradients.

Keep the int16 + scale_factor storage and the per-Frame texture cache;
only undo the R8 step. Worker now decodes int16 to Float32 ug/m³ and
posts a Float32Array; Pm25Layer uses r32float; shader restores pmMax.

+++ viewer/src/useForecast.ts
@@ -15,10 +15,7 @@ export type ForecastMeta = {
 };
 
 export type Frame = {
-  /** PM2.5 quantized to 0-255 against PM_MAX in the worker. Multiply by
-   *  PM_MAX/255 to recover µg/m³ for CPU-side consumers; the GPU samples
-   *  the r8unorm texture directly into the colormap LUT. */
-  data: Uint8Array;
+  data: Float32Array;
   maxPm25: number;
   initTime: number;
 };
